Task 3: AI Resume Screening System with Tracing/requirements.txt
langchain>=0.3.20
langchain-core>=0.3.45
langchain-openai>=0.3.7
langchain-huggingface>=0.1.2
langsmith>=0.3.8
langchain-groq>=0.2.4
python-dotenv>=1.0.1
pydantic>=2.10.6
